// libs/firefly-shared/src/state/composite/interests/interests.actions.enum.ts
export enum ActionsInterests {
  SetType = '[Interests] Set Type',
  SetVirtual = '[Interests] Set Virtual',
  SetSubscriptions = '[Interests] Set Subscriptions',
  Filter = '[Interests] Filter',
  Page = '[Interests] Page',

  SubscriptionToggle = '[Interests] Subscription Toggle',
  SubscriptionAdd = '[Interests] Subscription Add',
  SubscriptionRemove = '[Interests] Subscription Remove',
  SubscriptionOnOff = '[Interests] Subscription On/Off'
}